+++ src/plugin/PluginEditor.h
@@ -66,6 +66,7 @@ private:
     juce::Label aftertouchBendLabel;
     juce::Label neckSlideLabel;
     juce::Label slideFretStepsLabel;
+    juce::Label slideTailLabel;
     juce::Label lookaheadLabel;
     juce::Label fingerNoiseLabel;
     juce::Label ampFeedbackLabel;
@@ -100,6 +101,7 @@ private:
     juce::TextButton aftertouchBendInfoButton;
     juce::TextButton neckSlideInfoButton;
     juce::TextButton slideFretStepsInfoButton;
+    juce::TextButton slideTailInfoButton;
     juce::TextButton lookaheadInfoButton;
     juce::TextButton fingerNoiseInfoButton;
     juce::TextButton ampFeedbackInfoButton;
@@ -123,6 +125,7 @@ private:
     juce::TextButton palmMuteInfoButton;
     juce::TextButton harmonicTouchInfoButton;
     juce::ComboBox lookaheadBox;
+    juce::ComboBox slideTailBox;
     juce::ComboBox pickupModelBox;
     juce::ToggleButton vibratoModWheelSpeedButton;
     juce::ToggleButton vibratoModWheelDepthButton;
@@ -136,6 +139,7 @@ private:
     std::unique_ptr<SliderAttachment> aftertouchBendAttachment;
     std::unique_ptr<SliderAttachment> neckSlideAttachment;
     std::unique_ptr<SliderAttachment> slideFretStepsAttachment;
+    std::unique_ptr<ComboBoxAttachment> slideTailAttachment;
     std::unique_ptr<ComboBoxAttachment> lookaheadAttachment;
     std::unique_ptr<SliderAttachment> fingerNoiseAttachment;
     std::unique_ptr<SliderAttachment> ampFeedbackAttachment;